Moved helper components to helpers directory
