+++ widgets/deploymentWizardButtons/src/steps/helpers/ResourceAction.js
