+++ widgets/deploymentWizardButtons/src/steps/helpers/ResourceStatus.js
