+++ widgets/deploymentWizardButtons/src/steps/helpers/TaskList.js
